@@ -8,13 +8,13 @@
 	"author": "JustOptimize (Oggetto)",
 	"license": "GPL-3.0-only",
 	"devDependencies": {
-		"@babel/core": "^7.24.0",
-		"@babel/preset-env": "^7.24.0",
-		"@babel/preset-react": "^7.23.3",
+		"@babel/core": "^7.26.9",
+		"@babel/preset-env": "^7.26.9",
+		"@babel/preset-react": "^7.26.3",
 		"@biomejs/biome": "1.9.4",
-		"babel-loader": "^9.1.3",
+		"babel-loader": "^9.2.1",
 		"raw-loader": "^4.0.2",
-		"webpack": "^5.90.3",
+		"webpack": "^5.98.0",
 		"webpack-cli": "^5.1.4"
 	},
 	"babel": {
@@ -35,5 +35,8 @@
 				}
 			]
 		]
+	},
+	"dependencies": {
+		"@types/betterdiscord": "github:zerthox/betterdiscord-types"
 	}
 }
