@@ -1,4 +1,10 @@
-const TextElement = global.ZeresPluginLibrary?.DiscordModules?.TextElement;
+// @ts-check
+import { ModuleStore } from "../utils/modules";
+
+const {
+	Components: { TextElement },
+	React,
+} = ModuleStore;
 
 export default function ForumComponent({ channel }) {
 	if (channel.type !== 15) return null;
@@ -23,7 +29,7 @@ export default function ForumComponent({ channel }) {
 			Forum
 			{/* Tags */}
 			<TextElement
-				color={TextElement.Colors.INTERACTIVE_NORMAL}
+				color={TextElement.Colors.STANDARD}
 				size={TextElement.Sizes.SIZE_14}
 				style={{
 					marginTop: 24,
@@ -36,7 +42,7 @@ export default function ForumComponent({ channel }) {
 			{/* Guidelines */}
 			{channel.topic && (
 				<TextElement
-					color={TextElement.Colors.INTERACTIVE_NORMAL}
+					color={TextElement.Colors.STANDARD}
 					size={TextElement.Sizes.SIZE_14}
 					style={{
 						marginTop: 16,
@@ -47,7 +53,7 @@ export default function ForumComponent({ channel }) {
 			)}
 			{!channel.topic && (
 				<TextElement
-					color={TextElement.Colors.INTERACTIVE_NORMAL}
+					color={TextElement.Colors.STANDARD}
 					size={TextElement.Sizes.SIZE_14}
 					style={{
 						marginTop: 8,
